requirements.txt
chromadb
open-clip-torch
transformers
python-dotenv
flask
msgpack
mlx
numpy
torch
huggingface_hub
Pillow
git+https://github.com/harperreed/mlx_clip.git
clip